Cleaning code regarding tokens
