+++ spoti_stats/src/helper/redirectToAuthCodeFlow.js
@@ -4,13 +4,14 @@ import generateCodeVerifier from "./generateCodeVerifier";
 async function redirectToAuthCodeFlow(clientId) {
   const verifier = generateCodeVerifier(128);
   const challenge = await generateCodeChallenge(verifier);
+  const redirectUri = process.env.REDIRECT_URI ? process.env.REDIRECT_URI: "http://localhost:3000/callback" ;
 
   localStorage.setItem("verifier", verifier);
 
   const params = new URLSearchParams();
   params.append("client_id", clientId);
   params.append("response_type", "code");
-  params.append("redirect_uri", "http://localhost:3000/callback");
+  params.append("redirect_uri", redirectUri);
   //  params.append("redirect_uri", "https://spotistats-ai97.onrender.com/callback");
   params.append(
     "scope",
